Moved nanotdf code from gokas to sdk so it is more centrally located
