+++ sdk/nanotdfspec.ntdf
